requirements.txt
Flask~=2.2.2
azure-storage-blob==12.11.0
